nit(DBQ): Less doubles

+++ valhalla/baldr/double_bucket_queue.h
@@ -154,12 +154,12 @@ public:
   }
 
 private:
-  double bucketrange_; // Total range of costs in lower level buckets
-  double bucketsize_;  // Bucket size (range of costs in same bucket)
-  double inv_;         // 1/bucketsize (so we can avoid division)
+  float bucketrange_; // Total range of costs in lower level buckets
+  float bucketsize_;  // Bucket size (range of costs in same bucket)
+  float inv_;         // 1/bucketsize (so we can avoid division)
   double mincost_;     // Minimum cost within the low level buckets
-  double maxcost_;     // Above this goes into overflow bucket
-  double currentcost_; // Current cost
+  float maxcost_;     // Above this goes into overflow bucket
+  float currentcost_; // Current cost
 
   // Low level buckets
   buckets_t buckets_;
@@ -222,6 +222,7 @@ private:
         mincost_ += bucketrange_;
       }
       maxcost_ = mincost_ + bucketrange_;
+      //fprintf(stderr, "##    min %f mincost_ %f maxcost_ %f\n", min, mincost_, maxcost_);
 
       // Move elements within the range from overflow to buckets
       bucket_t tmp;
